requirements.txt
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.2
pycparser==2.22
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.2/libexec/wheel-0.45.1-py3-none-any.whl#sha256=b9235939e2096903717cb6bfc132267f8a7e46deb2ec3ef9c5e234ea301795d0
torch
torch-geometric==2.3.1
biopython
streamlit
web3
py-solc-x
eth-account
requests
pytz